final_k8s/orchestrator/requirements.txt
pika
loguru
cassandra-driver
